requirements.txt
torch
numpy
transformers
datasets
tiktoken
tqdm
regex
telethon
python-dotenv
asyncio
